requirements.txt
Django==2.2
djangorestframework==3.9.4
drf-yasg==1.15.0
coreapi==2.3.3
django-environ==0.4.5
django-extensions==2.1.3
django-model-utils==3.1.2
pytz==2018.5
mixer==6.1.3
flake8==3.7.7